Refactor, polish up prepanel serializer
