+++ talentmap_api/user_profile/urls/client.py
@@ -9,6 +9,6 @@ urlpatterns = [
     url(r'^(?P<pk>[0-9]+)/$', views.CdoClientView.as_view(get_retrieve), name='user_profile.UserProfile-client-detail'),
     url(r'^(?P<pk>[0-9]+)/survey/$', views.CdoClientSurveyView.as_view(get_list), name='bidding.StatusSurvey-client-list'),
     url(r'^(?P<pk>[0-9]+)/bids/$', views.CdoClientBidView.as_view(get_list), name='bidding.Bid-client-list'),
-    url(r'^(?P<pk>[0-9]+)/bids/(?P<bid_id>[0-9]+)/$', views.CdoClientBidView.as_view(get_retrieve), name='bidding.Bid-client-retrieve'),
+    url(r'^(?P<pk>[0-9]+)/bids/(?P<bid_id>[0-9]+)/prepanel/$', views.CdoClientBidView.as_view(get_retrieve), name='bidding.Bid-client-retrieve'),
     url(r'^(?P<pk>[0-9]+)/waivers/$', views.CdoClientWaiverView.as_view(get_list), name='bidding.Waiver-client-list'),
 ]
